requirements.txt
fastapi
uvicorn[standard]
opencv-python
langchain
langchain_openai
langchain_core
langchain_community
python-dotenv
pydantic
mediapipe
chromadb
facenet-pytorch
torch
torchvision
pillow